Added Defenders of Magic
I added the Heroes, Heroes II section, found we already had it, and then
updated the Heroes entry with the new book down the list. Deleted the
new entries, and created the Defenders of Magic entry. Which, on the
list is called Defending the Magic. I'm assuming it is the same thing.